models/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
asn1crypto=0.24.0=py36_0
blas=1.0=mkl
boto=2.49.0=py36_0
boto3=1.9.199=py_0
botocore=1.12.199=py_0
bz2file=0.98=py36_1
ca-certificates=2019.5.15=1
certifi=2019.6.16=py36_1
cffi=1.12.3=py36h2e261b9_0
chardet=3.0.4=py36_1003
cryptography=2.7=py36h1ba5d50_0
cryptography-vectors=2.3.1=py36_1000
cudatoolkit=9.2=0
docutils=0.14=py36_0
fasttext=0.2.0=hfc679d8_1
freetype=2.9.1=h8a8886c_1
gensim=3.4.0=py36h14c3975_0
idna=2.8=py36_0
intel-openmp=2019.4=243
jmespath=0.9.4=py_0
jpeg=9b=h024ee3a_2
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.0.10=h2733197_2
mkl=2019.4=243
mkl-service=2.0.2=py36h7b6447c_0
mkl_fft=1.0.14=py36ha843d7b_0
mkl_random=1.0.2=py36hd81dba3_0
ncurses=6.1=he6710b0_1
ninja=1.9.0=py36hfd86e86_0
numpy=1.16.4=py36h7e9f1db_0
numpy-base=1.16.4=py36hde5b4d6_0
olefile=0.46=py36_0
openssl=1.1.1=h7b6447c_0
pandas=0.25.0=py36he6710b0_0
pillow=6.1.0=py36h34e0f95_0
pip=19.1.1=py36_0
pycparser=2.19=py36_0
pyopenssl=19.0.0=py36_0
pysocks=1.7.0=py36_0
python=3.6.9=h265db76_0
python-dateutil=2.8.0=py36_0
pytorch=1.2.0=py3.6_cuda9.2.148_cudnn7.6.2_0
pytz=2019.2=py_0
readline=7.0=h7b6447c_5
requests=2.22.0=py36_0
s3transfer=0.2.1=py36_0
scipy=1.3.0=py36h7c811a0_0
setuptools=41.0.1=py36_0
six=1.12.0=py36_0
smart_open=1.8.4=py_0
sqlite=3.29.0=h7b6447c_0
tk=8.6.8=hbc83047_0
torchvision=0.4.0=py36_cu92
urllib3=1.24.2=py36_0
wheel=0.33.4=py36_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
